refactor: config 설정 변경 및 위치 변경
