--- src/main/java/com/example/global/AppConfig.java
@@ -1,14 +1,0 @@
-package com.example.global;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-@MapperScan("com.example")
-@ComponentScan("com.example")
-public class AppConfig implements WebMvcConfigurer {
-}
