+++ src/main/java/com/example/global/config/RootConfig.java
@@ -1,0 +1,18 @@
+package com.example.global.config;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RestController;
+
+@ComponentScan(
+        basePackages = "com.example",
+        excludeFilters = {
+                @ComponentScan.Filter(type = FilterType.ANNOTATION, classes = Controller.class),
+                @ComponentScan.Filter(type = FilterType.ANNOTATION, classes = RestController.class)
+        }
+)
+@MapperScan("com.example")
+public class RootConfig {
+}
