+++ src/main/resources/sql/init.sql
@@ -1,0 +1,6 @@
+DROP TABLE IF EXISTS user;
+
+CREATE TABLE user (
+                      id BIGINT AUTO_INCREMENT PRIMARY KEY,
+                      username VARCHAR(255) NOT NULL
+);
